headset/gyro/requirements.txt
paho-mqtt
smbus
